upgrade from cdk v1 to cdk v2
This change updates the requirements.txt to use the newer aws-cdk-lib
package and upgrades package versions for the necessary dependencies.

Based on those package changes, the imports are adjusted for both app.py
and the main stack application file. In addition to the import changes,
minor spacing changes were made.

The definition Construct Prop for the StateMachine has been deprecated.
This has been updated to use the new definition_body prop.